@@ -15,109 +15,106 @@
  */
 package org.springframework.sync;
 
-import static org.junit.Assert.*;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.Test;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.sync.json.JsonPatchPatchConverter;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Test;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.sync.json.JsonPatchPatchConverter;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
 
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
+class JsonPatchTest {
 
-public class JsonPatchTest {
+  @Test
+  void manySuccessfulOperations() throws Exception {
+    // initial Todo list
+    List<Todo> todos = new ArrayList<>();
+    todos.add(new Todo(1L, "A", true));
+    todos.add(new Todo(2L, "B", false));
+    todos.add(new Todo(3L, "C", false));
+    todos.add(new Todo(4L, "D", false));
+    todos.add(new Todo(5L, "E", false));
+    todos.add(new Todo(6L, "F", false));
 
-	@Test
-	public void manySuccessfulOperations() throws Exception {
-		// initial Todo list
-		List<Todo> todos = new ArrayList<Todo>();
-		todos.add(new Todo(1L, "A", true));
-		todos.add(new Todo(2L, "B", false));
-		todos.add(new Todo(3L, "C", false));
-		todos.add(new Todo(4L, "D", false));
-		todos.add(new Todo(5L, "E", false));
-		todos.add(new Todo(6L, "F", false));
-		
-		Patch patch = readJsonPatch("/org/springframework/sync/patch-many-successful-operations.json");
-		assertEquals(6, patch.size());
+    Patch patch = readJsonPatch("/org/springframework/sync/patch-many-successful-operations.json");
+    assertEquals(6, patch.size());
 
-		List<Todo> patchedTodos = patch.apply(todos, Todo.class);
-		
-		assertEquals(6, todos.size());
-		assertTrue(patchedTodos.get(1).isComplete());
-		assertEquals("C", patchedTodos.get(3).getDescription());
-		assertEquals("A", patchedTodos.get(4).getDescription());
-	}
+    List<Todo> patchedTodos = patch.apply(todos, Todo.class);
 
-	@Test
-	public void failureAtBeginning() throws Exception {
-		// initial Todo list
-		List<Todo> todos = new ArrayList<Todo>();
-		todos.add(new Todo(1L, "A", true));
-		todos.add(new Todo(2L, "B", false));
-		todos.add(new Todo(3L, "C", false));
-		todos.add(new Todo(4L, "D", false));
-		todos.add(new Todo(5L, "E", false));
-		todos.add(new Todo(6L, "F", false));
-		
-		Patch patch = readJsonPatch("/org/springframework/sync/patch-failing-operation-first.json");
+    assertEquals(6, todos.size());
+    assertTrue(patchedTodos.get(1).isComplete());
+    assertEquals("C", patchedTodos.get(3).getDescription());
+    assertEquals("A", patchedTodos.get(4).getDescription());
+  }
 
-		try {
-			patch.apply(todos, Todo.class);
-			fail();
-		} catch (PatchException e) {
-			assertEquals("Test against path '/5/description' failed.", e.getMessage());
-		}
-		
-		// nothing should have changed
-		assertEquals(6, todos.size());
-		assertFalse(todos.get(1).isComplete());
-		assertEquals("D", todos.get(3).getDescription());
-		assertEquals("E", todos.get(4).getDescription());
-		assertEquals("F", todos.get(5).getDescription());
-	}
-	
-	@Test
-	public void failureInMiddle() throws Exception {
-		// initial Todo list
-		List<Todo> todos = new ArrayList<Todo>();
-		todos.add(new Todo(1L, "A", true));
-		todos.add(new Todo(2L, "B", false));
-		todos.add(new Todo(3L, "C", false));
-		todos.add(new Todo(4L, "D", false));
-		todos.add(new Todo(5L, "E", false));
-		todos.add(new Todo(6L, "F", false));
-		
-		Patch patch = readJsonPatch("/org/springframework/sync/patch-failing-operation-in-middle.json");
+  @Test
+  void failureAtBeginning() throws Exception {
+    // initial Todo list
+    List<Todo> todos = new ArrayList<>();
+    todos.add(new Todo(1L, "A", true));
+    todos.add(new Todo(2L, "B", false));
+    todos.add(new Todo(3L, "C", false));
+    todos.add(new Todo(4L, "D", false));
+    todos.add(new Todo(5L, "E", false));
+    todos.add(new Todo(6L, "F", false));
 
-		try {
-			patch.apply(todos, Todo.class);
-			fail();
-		} catch (PatchException e) {
-			assertEquals("Test against path '/5/description' failed.", e.getMessage());
-		}
-		
-		// nothing should have changed
-		assertEquals(6, todos.size());
-		assertFalse(todos.get(1).isComplete());
-		assertEquals("D", todos.get(3).getDescription());
-		assertEquals("E", todos.get(4).getDescription());
-		assertEquals("F", todos.get(5).getDescription());
-	}
+    Patch patch = readJsonPatch("/org/springframework/sync/patch-failing-operation-first.json");
 
-	
-	
-	private Patch readJsonPatch(String jsonPatchFile) throws IOException, JsonParseException, JsonMappingException {
-		ClassPathResource resource = new ClassPathResource(jsonPatchFile);
-		ObjectMapper mapper = new ObjectMapper();
-		JsonNode node = mapper.readValue(resource.getInputStream(), JsonNode.class);
-		Patch patch = new JsonPatchPatchConverter().convert(node);
-		return patch;
-	}
-	
+    try {
+      patch.apply(todos, Todo.class);
+      fail();
+    } catch (PatchException e) {
+      assertEquals("Test against path '/5/description' failed.", e.getMessage());
+    }
+
+    // nothing should have changed
+    assertEquals(6, todos.size());
+    assertFalse(todos.get(1).isComplete());
+    assertEquals("D", todos.get(3).getDescription());
+    assertEquals("E", todos.get(4).getDescription());
+    assertEquals("F", todos.get(5).getDescription());
+  }
+
+  @Test
+  void failureInMiddle() throws Exception {
+    // initial Todo list
+    List<Todo> todos = new ArrayList<>();
+    todos.add(new Todo(1L, "A", true));
+    todos.add(new Todo(2L, "B", false));
+    todos.add(new Todo(3L, "C", false));
+    todos.add(new Todo(4L, "D", false));
+    todos.add(new Todo(5L, "E", false));
+    todos.add(new Todo(6L, "F", false));
+
+    Patch patch = readJsonPatch("/org/springframework/sync/patch-failing-operation-in-middle.json");
+
+    try {
+      patch.apply(todos, Todo.class);
+      fail();
+    } catch (PatchException e) {
+      assertEquals("Test against path '/5/description' failed.", e.getMessage());
+    }
+
+    // nothing should have changed
+    assertEquals(6, todos.size());
+    assertFalse(todos.get(1).isComplete());
+    assertEquals("D", todos.get(3).getDescription());
+    assertEquals("E", todos.get(4).getDescription());
+    assertEquals("F", todos.get(5).getDescription());
+  }
+
+  private Patch readJsonPatch(String jsonPatchFile) throws IOException {
+    ClassPathResource resource = new ClassPathResource(jsonPatchFile);
+    ObjectMapper mapper = new ObjectMapper();
+    JsonNode node = mapper.readValue(resource.getInputStream(), JsonNode.class);
+    return new JsonPatchPatchConverter().convert(node);
+  }
+
 }
