@@ -16,15 +16,16 @@
 
 package org.springframework.sync;
 
+
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+
 import java.io.Serializable;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-
-import org.apache.commons.lang.builder.EqualsBuilder;
-import org.apache.commons.lang.builder.HashCodeBuilder;
 
 /**
  * @author Roy Clarkson
@@ -33,61 +34,61 @@ import org.apache.commons.lang.builder.HashCodeBuilder;
 @Entity
 public class Todo implements Serializable {
 
-	private static final long serialVersionUID = 1L;
+  private static final long serialVersionUID = 1L;
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.AUTO)
+  private Long id;
 
-	private String description;
+  private String description;
 
-	private boolean complete;
+  private boolean complete;
 
-	public void setId(Long id) {
-		this.id = id;
-	}
-	
-	public Long getId() {
-		return id;
-	}
+  public void setId(Long id) {
+    this.id = id;
+  }
 
-	public String getDescription() {
-		return description;
-	}
-	
-	public void setDescription(String description) {
-		this.description = description;
-	}
+  public Long getId() {
+    return id;
+  }
 
-	public boolean isComplete() {
-		return complete;
-	}
-	
-	public void setComplete(boolean complete) {
-		this.complete = complete;
-	}
+  public String getDescription() {
+    return description;
+  }
 
-	public Todo() {
-	}
+  public void setDescription(String description) {
+    this.description = description;
+  }
 
-	public Todo(Long id, String description, boolean complete) {
-		this.id = id;
-		this.description = description;
-		this.complete = complete;
-	}
-	
-	@Override
-	public String toString() {
-		return "[ id=" + this.id + ", description=" + this.description + ", complete=" + this.complete + " ]";
-	}
+  public boolean isComplete() {
+    return complete;
+  }
 
-	@Override
-	public boolean equals(Object other) {
-		return EqualsBuilder.reflectionEquals(this, other);
-	}
-	
-	@Override
-	public int hashCode() {
-		return HashCodeBuilder.reflectionHashCode(this);
-	}
+  public void setComplete(boolean complete) {
+    this.complete = complete;
+  }
+
+  public Todo() {
+  }
+
+  public Todo(Long id, String description, boolean complete) {
+    this.id = id;
+    this.description = description;
+    this.complete = complete;
+  }
+
+  @Override
+  public String toString() {
+    return "[ id=" + this.id + ", description=" + this.description + ", complete=" + this.complete + " ]";
+  }
+
+  @Override
+  public boolean equals(Object other) {
+    return EqualsBuilder.reflectionEquals(this, other);
+  }
+
+  @Override
+  public int hashCode() {
+    return HashCodeBuilder.reflectionHashCode(this);
+  }
 }
